requirements.txt
# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Streamlit Dashboard
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0

# AI Enhancement (optional)
openai>=1.0.0
python-dotenv>=1.0.0

# PPT generation
python-pptx>=0.6.21
